pyproject.toml
[tool.poetry]
name = "a"
version = "0.1.0"
description = ""
authors = ["kesha1225 <samedov03@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
vkwave = {url = "https://github.com/fscdev/vkwave/archive/master.zip"}
pymorphy2 = "^0.9.1"
pytz = "^2021.1"
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
